Rename lockfiles to loaders
